fix: do not load binaries for CLI commands that don't need them
