+++ src/cli/commands/ChatCommand.ts
@@ -16,9 +16,9 @@ import {ReplHistory} from "../../utils/ReplHistory.js";
 import withStatusLogs from "../../utils/withStatusLogs.js";
 import {AlpacaChatWrapper} from "../../chatWrappers/AlpacaChatWrapper.js";
 import {FunctionaryChatWrapper} from "../../chatWrappers/FunctionaryChatWrapper.js";
-import {defineChatSessionFunction} from "../../llamaEvaluator/LlamaChatSession.js";
 import type {LlamaGrammar} from "../../llamaEvaluator/LlamaGrammar.js";
 import type {ModelTypeDescription} from "../../utils/getBin.js";
+import {defineChatSessionFunction} from "../../llamaEvaluator/LlamaChatSession/utils/defineChatSessionFunction.js";
 
 const modelWrappers = ["auto", "general", "llamaChat", "alpacaChat", "functionary", "chatML", "falconChat"] as const;
 
@@ -225,7 +225,7 @@ async function RunChat({
     jsonSchemaGrammarFile: jsonSchemaGrammarFilePath, threads, temperature, topK, topP, gpuLayers, lastTokensRepeatPenalty, repeatPenalty,
     penalizeRepeatingNewLine, repeatFrequencyPenalty, repeatPresencePenalty, maxTokens, noHistory, environmentFunctions
 }: ChatCommand) {
-    const {LlamaChatSession} = await import("../../llamaEvaluator/LlamaChatSession.js");
+    const {LlamaChatSession} = await import("../../llamaEvaluator/LlamaChatSession/LlamaChatSession.js");
     const {LlamaModel} = await import("../../llamaEvaluator/LlamaModel.js");
     const {LlamaContext} = await import("../../llamaEvaluator/LlamaContext/LlamaContext.js");
     const {LlamaGrammar} = await import("../../llamaEvaluator/LlamaGrammar.js");
