+++ src/llamaEvaluator/LlamaChatSession/utils/defineChatSessionFunction.ts
@@ -1,0 +1,18 @@
+import {GbnfJsonSchema, GbnfJsonSchemaToType} from "../../../utils/gbnfJson/types.js";
+import {ChatSessionModelFunction} from "../../../types.js";
+
+export function defineChatSessionFunction<const Params extends GbnfJsonSchema | undefined>({
+    description,
+    params,
+    handler
+}: {
+    description?: string,
+    params?: Params,
+    handler: (params: GbnfJsonSchemaToType<Params>) => any
+}): ChatSessionModelFunction<Params> {
+    return {
+        description,
+        params,
+        handler
+    };
+}
